requirements.txt
# niche-email-writer — requirements.txt

# Web App
streamlit>=1.30.0
streamlit-auth0>=0.3.0

# Backend API
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# AI
openai>=1.30.0

# Database / Auth
supabase>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic[email]>=2.7.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0
